requirements.txt
selenium==4.15.2
python-telegram-bot[job-queue]==20.7
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
schedule==1.2.0
